Copter: move prearm logic to arming_checks.cpp
